Simplify lemmy_context() function (dont return errors)
