+++ crates/apub/src/activities/post/create.rs
@@ -55,7 +55,7 @@ impl CreatePost {
       cc: [community.actor_id()],
       r#type: Default::default(),
       common: ActivityCommonFields {
-        context: lemmy_context()?.into(),
+        context: lemmy_context(),
         id: id.clone(),
         actor: actor.actor_id(),
         unparsed: Default::default(),
